chore: move property test skeleton out of FuzzPanel
