+++ src/ui/FuzzPanel.ts
@@ -15,14 +15,7 @@ import {
 } from "./CoverageHeatmap";
 import { normalizePathForKey } from "../fuzzer/Util";
 import { CodeCoverageMeasureStats } from "../fuzzer/measures/CoverageMeasure";
-
-// Consts for validator result arg name generation
-const resultArgCandidateNames = ["r", "result", "_r", "_result"];
-const maxResultArgSuffix = 1000;
-
-// Consts for validator out variable name generation
-const outVarCandidateNames = ["out", "output", "_out", "_output"];
-const maxOutVarSuffix = 1000;
+import { getPropertyTestSkeleton } from "../fuzzer/analysis/typescript/Util";
 
 /**
  * FuzzPanel displays fuzzer options, actions, and the last results for a
@@ -927,43 +920,11 @@ export class FuzzPanel {
     const hasImport = Object.keys(program.getImports()).some(
       (e) => e === "FuzzTestResult"
     );
-
-    const inArgs = fn.getArgDefs();
-    const validatorArgs = this._getValidatorArgs(inArgs);
-    const inArgConsts = inArgs
-      .map(
-        (argDef, i) =>
-          `  const ${argDef.getName()}: ${argDef.getTypeAnnotation()} = ${
-            validatorArgs.resultArgName
-          }.in[${i}];`
-      )
-      .join("\n");
-
-    const outTypeAsArg = fn.getReturnArg();
-    const outTypeAsString = outTypeAsArg
-      ? outTypeAsArg.getTypeAnnotation()
-      : undefined;
-
-    const outArgConst = this._getOutArgConst(
-      inArgs,
-      validatorArgs.resultArgName,
-      outTypeAsString
-    );
-
-    // Name of the validator generated
-    const validatorName = `${validatorPrefix}${
-      fnCounter === 0 ? "" : fnCounter
-    }`;
-
-    // prettier-ignore
+    const validatorSuffix = fnCounter === 0 ? "" : fnCounter.toString();
+    const validatorName = `${validatorPrefix}${validatorSuffix}`;
     const skeleton = `
-
-export function ${validatorName}${validatorArgs.str}: "pass" | "fail" | "unknown" {
-${inArgConsts}
-  ${outArgConst}
-
-  return "pass";
-}`;
+    
+${getPropertyTestSkeleton(this._fuzzEnv.function, validatorSuffix)}`;
 
     // Save the editor if dirty
     for (const editor of vscode.window.visibleTextEditors) {
@@ -1016,102 +977,6 @@ ${inArgConsts}
       );
     }
   }
-
-  /**
-   * Choose a name for an identifier that doesn't conflict with the input arguments
-   *
-   * @param inArgs The input arguments
-   * @param candidateNames The candidate names to choose from
-   * @param maxSuffix The maximum suffix to use when generating a new name
-   * @returns The chosen name and whether it was generated
-   */
-  private _getIdentifierNameAvoidingConflicts(
-    // The input arguments
-    inArgs: fuzzer.ArgDef<fuzzer.ArgType>[],
-    // The candidate names to choose from
-    candidateNames: string[],
-    // The maximum suffix to use when generating a new name
-    maxSuffix: number
-  ): {
-    // The chosen name
-    name: string;
-    // Whether the name was generated (as opposed to being in possibleResultArgNames)
-    generated: boolean;
-  } {
-    const inArgNames = inArgs.map((argDef) => argDef.getName());
-    for (const name of candidateNames) {
-      if (!inArgNames.includes(name)) {
-        return { name, generated: false };
-      }
-    }
-
-    let i = 1;
-    // Generate a new name with a suffix
-    for (const candidateName of candidateNames) {
-      while (i <= maxSuffix) {
-        const name = `${candidateName}_${i}`;
-        if (!inArgNames.includes(name)) {
-          return { name, generated: true };
-        }
-        i++;
-      }
-    }
-
-    // In the extremely unlikely event that all the names generated above are
-    // already in `inArgNames`, we'll just return `r_conflicted` and not worry
-    // about potential conflicts.
-    return { name: "r_conflicted", generated: true };
-  } // fn: getIdentifierNameAvoidingConflicts()
-
-  /**
-   * Get the string representation for the validator arguments, along with the
-   * name of the argument that will hold the result.
-   *
-   * @param inArgs The input arguments
-   * @returns An object containing the above information
-   */
-  private _getValidatorArgs(inArgs: fuzzer.ArgDef<fuzzer.ArgType>[]): {
-    str: string;
-    resultArgName: string;
-  } {
-    const resultArgName = this._getIdentifierNameAvoidingConflicts(
-      inArgs,
-      resultArgCandidateNames,
-      maxResultArgSuffix
-    );
-    const resultArgString = `${resultArgName.name}: FuzzTestResult`;
-    return {
-      str: `(${resultArgString})`,
-      resultArgName: resultArgName.name,
-    };
-  } // fn: getValidatorArgs()
-
-  /**
-   * Get the string for the declaration of the out variable.
-   *
-   * The out variable is the variable that will hold the result of the function
-   * under test.
-   *
-   * @param inArgs The input arguments
-   * @param resultArgName The name of the argument that will hold the result
-   * @param returnType The return type of the function
-   * @returns The string for the declaration of the out variable
-   */
-  private _getOutArgConst(
-    inArgs: fuzzer.ArgDef<fuzzer.ArgType>[],
-    resultArgName: string,
-    returnType?: string
-  ): string {
-    const outVarName = this._getIdentifierNameAvoidingConflicts(
-      inArgs,
-      outVarCandidateNames,
-      maxOutVarSuffix
-    );
-    const outVarString = `const ${outVarName.name}${
-      returnType ? ": " + returnType : ""
-    } = ${resultArgName}.out;`;
-    return outVarString;
-  } // fn: getOutConst()
 
   /**
    * Message handler for the `validator.getList` command. Gets the list
